Align test classes and structure
