+++ libs/ui-lib-tests/cypress/views/forms/Operators/OperatorsForm.ts
